semana-04/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
kagglehub>=0.2.0
anthropic>=0.30.0